fix(transport): X.01 — FaceId monotonic u64, no recycling (closes ABA)
`FaceTable::alloc_id()` previously handed out u32 ids and pushed
removed ids onto a free list — a classic ABA hazard for any code
that stamps a face id and dereferences it later (e.g. NDNLPv2
`IncomingFaceId`, a `source_face` captured before an await, a tag
on a `PacketContext`):

  face X (id N) closes → N pushed onto free list →
  face Y allocated and receives N →
  stamped "N" now refers to Y, not X.

Worst case: mgmt rib/register binds the wrong face — an
authorization-boundary bug — even with the E.02 LP-`IncomingFaceId`
path doing the right thing on its end.

Fix mirrors NFD (`daemon/fw/face-table.cpp:58` `++m_lastFaceId` on
`uint64_t`): `FaceId(pub u64)`, `alloc_id()` is `fetch_add` on an
`AtomicU64`, `remove()` drops the face but does not free-list the
id.  Reserved-range carve-out (`0xFFFF_0000`) deleted; the demo CA's
hand-rolled high-range face id replaced with `engine.faces().alloc_id()`.

PIT InRecord/OutRecord widened to `u64` face_id to match.  Wire
formats are already `NonNegativeInteger` (any u64), so no protocol
changes.  WASM strategy ABI continues to use u32 over the
host/guest boundary with a TODO — practical face counters don't
exceed `u32::MAX` in any realistic deployment.

Witness:  testbed/tests/audit/x01_face_id_monotonic.sh (GREP + RUST-UNIT).
Audit:    docs/notes/spec-compliance-audit-2026-04-20.md (E.02 entry).
Lessons:  memory/feedback_face_id_no_recycle.md (new),
          memory/feedback_face_id_provenance.md (updated).

+++ binaries/tooling/ndn-bench/src/main.rs
@@ -107,7 +107,7 @@ async fn main() -> Result<()> {
             // Measure the overhead of the InProcFace channel send path.
             // InProcFace::new creates (engine-side face, app-side handle).
             // The "engine" receives via face.recv(); the "app" sends via handle.
-            let (face, handle) = InProcFace::new(FaceId(worker as u32), 128);
+            let (face, handle) = InProcFace::new(FaceId(worker), 128);
             use ndn_transport::Face;
             let _face_id = face.id(); // confirm the face is alive
